Add project list component and refactor CSS folder
Every modal component will have to import
modal.css in order to has the mutual styles for
modals (specifically project list modal in this
commit).

Now the CSS folder (styles), specifically main.css
is splitted into smaller files with specific
functions and roles.

Change the primary color of the application to
blue (it is simpler to combine with other colors)