Add todo detail UI component
The todo detail section appeared next to the main
todo list when clicking one of the todo items. It
has a title, a description, a due date and
optionally a checklist.

Besides, remove edit button in todo item. Now when
the user clicks the todo item itself, they
activate it.

Add join() after map() in order to fix the issue
when being rendered into view, the items are
separated by commas.

In button.css, .btn-square class is used for
buttons with only icon and without text.