Add and fix small UI details to increase UX
Add the rotating animation to the plus icon of the
new task button.

In Checklist part of todoDetail, expand the label
to take up all space of the parent li element. Now
when the user hovers on everything inside the li
element, they will be able to click to set checked
to that checklist item.