Add projectListModal toggle and refactor component
Now the user is able to click the menu button at
the top-right corner of the app to open the
current project list.

Refactor Header and ProjectListModal from factory
functions to classes because I want to add
inheritance to the component system. Specifically,
now every component which is a modal will inherit
from the Modal component. Moreover, classes allow
adding methods, which is useful for component's
event handling. (I think they are cleaner than
factory functions in terms of adding methods)

In projectListModal.css, add a specified class
to the selector as well as the component itself in
order to override the style in modal.css
(Following CSS specificity rules).

In modal.css, slightly modify the zoomOut
animation of the modal so that it is faster and
smoother.