Add editing project's title and desc feature
The user is able to click the pen next to the
title and the description to change their content.

Separate the resetContent function into
resetHeader and resetTodoListSection and put them
at first line of the corresponding rendering
functions for easier DOM manipulation.

Fix the 'element' reference in editTodoModal by
adding 'this' (because it is in a class).

Set buttons' and inputs' color as 'inherit' so
that the title and description input will have the
color of the text itself.